refactor: Filters
